Changed Encapsulation Issue
Changed UpdateFile to Public, The Details of an Item will now be an array of strings as a list allows anyone to edit the details list inside each item, furthermore, the UpdateItem Function in the ItemCollection Class is removed, where changes to an item will be done directly on the item using setDetails. This allows Item to be edited from anywhere, by obtaining it using filter, getrelateditems, etc. In addition, Since the Relationship between PR and Product and Sales and Product requires a quantity, A factory is used to manage the creation of PR and Sales as the original method of creating a relationship requires the change of value in the QUANTITY field of an item which if not done properly could lead to the loss of data integrity